+++ frontend/app/routes/_gcweb-app.personal-information.home-address.edit.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 
 import { type ActionFunctionArgs, type LoaderFunctionArgs, json, redirect } from '@remix-run/node';
 import { Form, Link, useActionData, useLoaderData } from '@remix-run/react';
@@ -17,26 +17,25 @@ const i18nNamespaces = getTypedI18nNamespaces('personal-information');
 
 export const handle = {
   breadcrumbs: [
-    { labelI18nKey: 'personal-information:mailing-address.edit.breadcrumbs.home', to: '/' },
-    { labelI18nKey: 'personal-information:mailing-address.edit.breadcrumbs.personal-information', to: '/personal-information' },
-    { labelI18nKey: 'personal-information:mailing-address.edit.breadcrumbs.mailing-address-change' },
+    { labelI18nKey: 'personal-information:home-address.edit.breadcrumbs.home', to: '/' },
+    { labelI18nKey: 'personal-information:home-address.edit.breadcrumbs.personal-information', to: '/personal-information' },
+    { labelI18nKey: 'personal-information:home-address.edit.breadcrumbs.home-address-change' },
   ],
   i18nNamespaces,
   pageIdentifier: 'CDCP-0004',
-  pageTitleI18nKey: 'personal-information:mailing-address.edit.page-title',
+  pageTitleI18nKey: 'personal-information:home-address.edit.page-title',
 } as const satisfies RouteHandleData;
 
 export async function loader({ request }: LoaderFunctionArgs) {
   const userId = await userService.getUserId();
   const userInfo = await userService.getUserInfo(userId);
-  const addressInfo = await addressService.getAddressInfo(userId, userInfo?.mailingAddress ?? '');
-  const homeAddressInfo = await addressService.getAddressInfo(userId, userInfo?.homeAddress ?? '');
+  const addressInfo = await addressService.getAddressInfo(userId, userInfo?.homeAddress ?? '');
 
   if (!userInfo) {
     throw new Response(null, { status: 404 });
   }
 
-  return json({ addressInfo, homeAddressInfo });
+  return json({ addressInfo });
 }
 
 export async function action({ request }: ActionFunctionArgs) {
@@ -68,34 +67,28 @@ export async function action({ request }: ActionFunctionArgs) {
   }
 
   const session = await sessionService.getSession(request.headers.get('Cookie'));
-  session.set('newMailingAddress', parsedDataResult.data);
+  session.set('newHomeAddress', parsedDataResult.data);
 
-  return redirect('/personal-information/mailing-address/confirm', {
+  return redirect('/personal-information/home-address/confirm', {
     headers: {
       'Set-Cookie': await sessionService.commitSession(session),
     },
   });
 }
 
-export default function PersonalInformationMailingAddressEdit() {
+export default function PersonalInformationHomeAddressEdit() {
   const actionData = useActionData<typeof action>();
-  const { addressInfo, homeAddressInfo } = useLoaderData<typeof loader>();
+  const { addressInfo } = useLoaderData<typeof loader>();
   const { t } = useTranslation(i18nNamespaces);
   const errorSummaryId = 'error-summary';
-  const [isCopyAddressChecked, setCopyAddressChecked] = useState(false);
-
-  const checkHandler = () => {
-    setCopyAddressChecked((curState) => !curState);
-  };
 
   const defaultValues = {
-    address: actionData?.formData.address ?? (isCopyAddressChecked ? homeAddressInfo?.address : addressInfo?.address),
-    city: actionData?.formData.city ?? (isCopyAddressChecked ? homeAddressInfo?.city : addressInfo?.city),
-    province: actionData?.formData.province ?? (isCopyAddressChecked ? homeAddressInfo?.province : addressInfo?.province),
-    country: actionData?.formData.country ?? (isCopyAddressChecked ? homeAddressInfo?.country : addressInfo?.country),
-    postalCode: actionData?.formData.postalCode ?? (isCopyAddressChecked ? homeAddressInfo?.postalCode : addressInfo?.postalCode),
+    address: actionData?.formData.address ?? addressInfo?.address,
+    city: actionData?.formData.city ?? addressInfo?.city,
+    province: actionData?.formData.province ?? addressInfo?.province,
+    country: actionData?.formData.country ?? addressInfo?.country,
+    postalCode: actionData?.formData.postalCode ?? addressInfo?.postalCode,
   };
-
   /**
    * Gets an error message based on the provided internationalization (i18n) key.
    *
@@ -109,7 +102,7 @@ export default function PersonalInformationMailingAddressEdit() {
      * The 'as any' is employed to circumvent typechecking, as the type of
      * 'errorI18nKey' is a string, and the string literal cannot undergo validation.
      */
-    return t(`personal-information:mailing-address.edit.error-message.${errorI18nKey}` as any);
+    return t(`personal-information:home-address.edit.error-message.${errorI18nKey}` as any);
   }
 
   const errorMessages = {
@@ -132,63 +125,18 @@ export default function PersonalInformationMailingAddressEdit() {
     <>
       {errorSummaryItems.length > 0 && <ErrorSummary id={errorSummaryId} errors={errorSummaryItems} />}
       <Form className="max-w-prose" method="post">
-        <div className="checkbox gc-chckbxrdio">
-          <input id="copy-home-address" type="checkbox" name="ifCopyHomeAddress" checked={isCopyAddressChecked} onChange={checkHandler} />
-          <label id="copy-home-address" htmlFor="copy-home-address">
-            Copy home address
-          </label>
-        </div>
-
-        <InputField
-          id="address"
-          label={t('personal-information:mailing-address.edit.field.address')}
-          name="address"
-          required
-          key={isCopyAddressChecked ? homeAddressInfo?.address : addressInfo?.address}
-          defaultValue={defaultValues.address}
-          errorMessage={errorMessages.address}
-        />
-        <InputField
-          id="city"
-          label={t('personal-information:mailing-address.edit.field.city')}
-          name="city"
-          required
-          key={isCopyAddressChecked ? homeAddressInfo?.city : addressInfo?.city}
-          defaultValue={defaultValues.city}
-          errorMessage={errorMessages.city}
-        />
-        <InputField
-          id="province"
-          label={t('personal-information:mailing-address.edit.field.province')}
-          name="province"
-          key={isCopyAddressChecked ? homeAddressInfo?.province : addressInfo?.province}
-          defaultValue={defaultValues.province}
-          errorMessage={errorMessages.province}
-        />
-        <InputField
-          id="postalCode"
-          label={t('personal-information:mailing-address.edit.field.postal-code')}
-          name="postalCode"
-          key={isCopyAddressChecked ? homeAddressInfo?.postalCode : addressInfo?.postalCode}
-          defaultValue={defaultValues.postalCode}
-          errorMessage={errorMessages.postalCode}
-        />
-        <InputField
-          id="country"
-          label={t('personal-information:mailing-address.edit.field.country')}
-          name="country"
-          required
-          key={isCopyAddressChecked ? homeAddressInfo?.country : addressInfo?.country}
-          defaultValue={defaultValues.country}
-          errorMessage={errorMessages.country}
-        />
+        <InputField id="address" label={t('personal-information:home-address.edit.field.address')} name="address" required defaultValue={defaultValues.address} errorMessage={errorMessages.address} />
+        <InputField id="city" label={t('personal-information:home-address.edit.field.city')} name="city" required defaultValue={defaultValues.city} errorMessage={errorMessages.city} />
+        <InputField id="province" label={t('personal-information:home-address.edit.field.province')} name="province" defaultValue={defaultValues.province} errorMessage={errorMessages.province} />
+        <InputField id="postalCode" label={t('personal-information:home-address.edit.field.postal-code')} name="postalCode" defaultValue={defaultValues.postalCode} errorMessage={errorMessages.postalCode} />
+        <InputField id="country" label={t('personal-information:home-address.edit.field.country')} name="country" required defaultValue={defaultValues.country} errorMessage={errorMessages.country} />
 
         <div className="flex flex-wrap gap-3">
           <button id="change-button" className="btn btn-primary btn-lg">
-            {t('personal-information:mailing-address.edit.button.change')}
+            {t('personal-information:home-address.edit.button.change')}
           </button>
           <Link id="cancel-button" to="/personal-information" className="btn btn-default btn-lg">
-            {t('personal-information:mailing-address.edit.button.cancel')}
+            {t('personal-information:home-address.edit.button.cancel')}
           </Link>
         </div>
       </Form>
